added a helpers.rb to house all the global functions and completed login auth
